backend/requirements.txt
fastapi
uvicorn
requests
paho-mqtt
psycopg2-binary
